Update kicad packages to 5.1.4. Changes since 5.1.2:
The KiCad project is proud to announce the latest series 5 stable release of
KiCad. The 5.1.4 stable version contains critical bug fixes and other minor
improvements since the 5.1.2 and 5.1.3 releases. It also includes improved
footprint, symbol, and 3D model libraries, translations, and documentation.
A list of all of the fixed bugs since the 5.1.2 and 5.1.3 releases can be
found on the KiCad 5.1.3 milestone page and the KiCad 5.1.4 milestone page.
This release contains several critical bug fixes so please consider upgrading
as soon as possible.