lang/ruby24: update to 2.4.9
Update ruby24-base and related packges to 2.4.9.

pkgsrc chagnes

* fix warnings of pkglint.

Quote from release announce:

Ruby 2.4.8 (2019-10-01)

This release includes security fixes. Please check the topics below for
details.

* CVE-2019-16255: A code injection vulnerability of Shell#[] and Shell#test
* CVE-2019-16254: HTTP response splitting in WEBrick (Additional fix)
* CVE-2019-15845: A NUL injection vulnerability of File.fnmatch and
  File.fnmatch?
* CVE-2019-16201: Regular Expression Denial of Service vulnerability of
  WEBrick¡Çs Digest access authentication


Ruby 2.4.9 (2019-10-02)

This release is a re-package of 2.4.8 because the previous Ruby 2.4.8
release tarball does not install. (See [Bug #16197] in detail.) There are no
essential change except their version numbers between 2.4.8 and 2.4.9.

Ruby 2.4 is now under the state of the security maintenance phase, until the
end of March of 2020. After that date, maintenance of Ruby 2.4 will be
ended. We recommend you start planning the migration to newer versions of
Ruby, such as 2.6 or 2.5.