editors/xournal: add patch to smooth input coordinates
On NetBSD, absolute coordinate changes are reported as an X value
followed by a Y value, which (when moving quickly) leads to a
"stepping" effect. Using a Bezier curve fitting for the points
yields a smooth curve in most cases, and does not seem to adversely
affect the user experience (even with a conventional mouse).